Form/List: removed the ListInfo_t struct
Change its attributes to WndListFrame properties.  Make all of these unsigned.

+++ Common/Source/Dialogs/dlgAirspaceWarning.cpp
@@ -313,7 +313,7 @@ OnAirspaceListItemPaint(Canvas &canvas, const RECT paint_rc, unsigned i)
 
     if (!AirspaceWarnGetItem(i, pAS)) return;
 
-    if ((int)i == wAirspaceList->GetCursorIndex())
+    if (i == wAirspaceList->GetCursorIndex())
       FocusedID = pAS.ID;
 
     if (pAS.IsCircle){
